Fix "String.Format" helper to handle generic formatting
"String.Format" used to only support "DateTime" for formatting.

As for now, as soon as the value implements "IFormattable" that interface is used to perform the formatting. Also, if the format provider exposes a "ICustomFormatter" service, that formatter is used first for formatting the value.

The formatting falls back to "ToString()" or to an empty string if the value to format is null.